move fixtures to fixtures.py and fix a few warnings
