+++ napari_animation/_tests/test_animation.py
@@ -5,42 +5,7 @@ import pytest
 
 from napari_animation import Animation
 
-
-@pytest.fixture
-def make_napari_viewer():
-    from napari import Viewer
-    viewers = []
-
-    def actual_factory(*model_args, viewer_class=Viewer, **model_kwargs):
-        model_kwargs.setdefault('show', False)
-        viewer = viewer_class(*model_args, **model_kwargs)
-        viewers.append(viewer)
-        return viewer
-
-    yield actual_factory
-
-    for viewer in viewers:
-        viewer.close()
-
-
-@pytest.fixture
-def empty_animation(make_napari_viewer):
-    viewer = make_napari_viewer()
-    animation = Animation(viewer)
-    return animation
-
-
-@pytest.fixture
-def animation_with_keyframes(empty_animation):
-    for i in range(2):
-        empty_animation.capture_keyframe()
-        empty_animation.viewer.camera.zoom *= 2
-    return empty_animation
-
-
-@pytest.fixture
-def viewer_state(empty_animation):
-    return empty_animation._get_viewer_state()
+from .fixtures import make_napari_viewer, empty_animation, animation_with_keyframes, viewer_state
 
 
 def test_animation(make_napari_viewer):
@@ -84,8 +49,8 @@ def test_set_viewer_state(animation_with_keyframes, viewer_state):
     current_state = animation._get_viewer_state()
     animation._set_viewer_state(viewer_state)
 
-    animation_dims_state = animation.viewer.dims.asdict()
-    animation_camera_state = animation.viewer.camera.asdict()
+    animation_dims_state = animation.viewer.dims.dict()
+    animation_camera_state = animation.viewer.camera.dict()
 
     assert animation_dims_state == current_state['dims']
     for key in ('center', 'angles', 'interactive'):
